+++ src/main/java/net/geforcemods/securitycraft/blockentities/PortableRadarBlockEntity.java
@@ -1,6 +1,5 @@
 package net.geforcemods.securitycraft.blockentities;
 
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
@@ -61,14 +60,7 @@ public class PortableRadarBlockEntity extends CustomizableBlockEntity implements
 				PortableRadarBlock.togglePowerOutput(level, pos, !closebyPlayers.isEmpty());
 
 			if (!closebyPlayers.isEmpty()) {
-				Collection<ServerPlayer> onlineTeamPlayers = TeamUtils.getOnlinePlayersInTeam(level.getServer(), getOwner());
-
-				if (onlineTeamPlayers.isEmpty()) { //owner may not be in a team
-					ServerPlayer ownerPlayer = level.getServer().getPlayerList().getPlayerByName(getOwner().getName());
-
-					if (ownerPlayer != null)
-						onlineTeamPlayers = Arrays.asList(ownerPlayer);
-				}
+				Collection<ServerPlayer> onlineTeamPlayers = TeamUtils.getOnlinePlayersFromOwner(level.getServer(), getOwner());
 
 				for (Player closebyPlayer : closebyPlayers) {
 					if (shouldSendMessage(closebyPlayer)) {