+++ src/main/java/net/geforcemods/securitycraft/entity/SecuritySeaBoat.java
@@ -394,7 +394,7 @@ public class SecuritySeaBoat extends ChestBoat implements IOwnable, IPasscodePro
 	}
 
 	@Override
-	public void onOwnerChanged(BlockState state, Level level, BlockPos pos, Player player) {}
+	public void onOwnerChanged(BlockState state, Level level, BlockPos pos, Player player, Owner oldOwner, Owner newOwner) {}
 
 	@Override
 	public void activate(Player player) {